fix(debt): retire Tier 0 defects and Tier 1 quick wins
Tier 0 — latent defects:
- JNI: bound the output direct buffer with GetDirectBufferCapacity and
  thread the capacity into run_full_field so a large ROI/step can no
  longer overflow it; null-check every GetByteArrayElements.
- Remove five dead engine toggles (use_delaunay/fallback/rgdic, blur,
  nlvc) end-to-end across Kotlin <-> JNI <-> pipeline, and delete the
  entire NLVC strain path (VSG is now the only strain method).
- Real release signing: signingConfigs.release reads the SIGNING_* env
  vars; release.yml verifies the APK signature with apksigner and fails
  if it is not release-signed; version driven by workflow inputs.

Tier 1 — quick wins:
- CI: replace the broken pull_request.changed_files gate and the dead
  skip-native expression with a dorny/paths-filter changes job so the
  backend tier actually runs on PRs.
- Native: -Wall -Wextra on first-party targets (vendored headers SYSTEM);
  fix two sign-compare warnings; opt-in SEMPER_WERROR verified clean on
  the host path.
- Backend: fail startup on missing required env vars when deployed, parse
  ints with a clear error (+ tests); harden GET /v1/me/export with
  verified_device; sanitize the Content-Disposition filename.
- Android: StaticAnalysisActivity.onDestroy cancels the native engine,
  clears KEEP_SCREEN_ON, and releases the overlay ticker.
- Solver clears the process-global cancel flag on entry.
- Coverage (report-only): Kover, pytest-cov, native --coverage.
- Housekeeping: delete 3 tautological tests + dead code (ENGINE_ERROR_*,
  Image::gradient_x/y, drive.stream_file), gitignore keystores/coverage,
  drop 19 orphaned readme JPEGs, correct stale test/release docs.

GridPoint hardened via default member initializers (C++17 has no
designated initializers; full conversion deferred to the tested Phase-3
work on full_field_solver). /v1/files/{id}/content left on current_user
deliberately — it is the app restore path and needs coordinated client
signing.

Verified: native host tests 48 pass under -Werror; backend 20 pass, ruff
clean; Kotlin compiles + unit tests + spotless + detekt pass. The Android
NDK build was not compilable on the dev host.

+++ native/src/math/optimization_engine.cpp
@@ -205,7 +205,7 @@ namespace Semper {
             // 🚀 FAST-PATH: All pixels are safely inside the image boundaries.
             // Portable SIMD kernels — one codepath compiled to NEON on ARM
             // and SSE on x86, replacing the old #if __aarch64__ fork.
-            if (valid_pixels == n) {
+            if (valid_pixels == static_cast<int>(n)) {
                 def_sum_sq = simd::sum_sq_diff(def_vals.data(), n, def_mean);
 
                 float def_std = std::sqrt(def_sum_sq / valid_pixels);
@@ -370,7 +370,7 @@ namespace Semper {
         float def_sum_sq = 0.0f;
         float znssd = 0.0f;
 
-        if (valid_pixels == n) {
+        if (valid_pixels == static_cast<int>(n)) {
             // 🚀 Portable SIMD kernels (NEON/SSE via one codepath) — see SimdKernels.h
             def_sum_sq = simd::sum_sq_diff(buffer.data(), n, def_mean);
